move paging tests in a single file
